--- genesyscloud/webdeployments_configuration/resource_genesyscloud_webdeployments_configuration_utils.go
@@ -1,834 +1,0 @@
-package webdeployments_configuration
-
-import (
-	"context"
-	"strings"
-	"terraform-provider-genesyscloud/genesyscloud/util/lists"
-	"terraform-provider-genesyscloud/genesyscloud/util/resourcedata"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/mypurecloud/platform-client-sdk-go/v119/platformclientv2"
-)
-
-func customizeConfigurationDiff(ctx context.Context, diff *schema.ResourceDiff, meta interface{}) error {
-	if len(diff.GetChangedKeysPrefix("")) > 0 {
-		// When any change is made to the configuration we automatically publish a new version, so mark the version as updated
-		// so dependent deployments will update appropriately to reference the newest version
-		diff.SetNewComputed("version")
-	}
-	return nil
-}
-
-func readSelectorEventTriggers(triggers []interface{}) *[]platformclientv2.Selectoreventtrigger {
-	if triggers == nil || len(triggers) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Selectoreventtrigger, len(triggers))
-	for i, value := range triggers {
-		if trigger, ok := value.(map[string]interface{}); ok {
-			selector := trigger["selector"].(string)
-			eventName := trigger["event_name"].(string)
-			results[i] = platformclientv2.Selectoreventtrigger{
-				Selector:  &selector,
-				EventName: &eventName,
-			}
-		}
-	}
-
-	return &results
-}
-
-func readFormsTrackTriggers(triggers []interface{}) *[]platformclientv2.Formstracktrigger {
-	if triggers == nil || len(triggers) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Formstracktrigger, len(triggers))
-	for i, value := range triggers {
-		if trigger, ok := value.(map[string]interface{}); ok {
-			selector := trigger["selector"].(string)
-			formName := trigger["form_name"].(string)
-			captureDataOnAbandon := trigger["capture_data_on_form_abandon"].(bool)
-			captureDataOnSubmit := trigger["capture_data_on_form_submit"].(bool)
-			results[i] = platformclientv2.Formstracktrigger{
-				Selector:                 &selector,
-				FormName:                 &formName,
-				CaptureDataOnFormAbandon: &captureDataOnAbandon,
-				CaptureDataOnFormSubmit:  &captureDataOnSubmit,
-			}
-		}
-	}
-
-	return &results
-}
-
-func readIdleEventTriggers(triggers []interface{}) *[]platformclientv2.Idleeventtrigger {
-	if triggers == nil || len(triggers) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Idleeventtrigger, len(triggers))
-	for i, value := range triggers {
-		if trigger, ok := value.(map[string]interface{}); ok {
-			eventName := trigger["event_name"].(string)
-			idleAfterSeconds := trigger["idle_after_seconds"].(int)
-			results[i] = platformclientv2.Idleeventtrigger{
-				EventName:        &eventName,
-				IdleAfterSeconds: &idleAfterSeconds,
-			}
-		}
-	}
-
-	return &results
-}
-
-func readScrollPercentageEventTriggers(triggers []interface{}) *[]platformclientv2.Scrollpercentageeventtrigger {
-	if triggers == nil || len(triggers) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Scrollpercentageeventtrigger, len(triggers))
-	for i, value := range triggers {
-		if trigger, ok := value.(map[string]interface{}); ok {
-			eventName := trigger["event_name"].(string)
-			percentage := trigger["percentage"].(int)
-			results[i] = platformclientv2.Scrollpercentageeventtrigger{
-				EventName:  &eventName,
-				Percentage: &percentage,
-			}
-		}
-	}
-
-	return &results
-}
-
-func readCustomMessages(messages []interface{}) *[]platformclientv2.Supportcentercustommessage {
-	if messages == nil || len(messages) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Supportcentercustommessage, len(messages))
-	for i, value := range messages {
-		if message, ok := value.(map[string]interface{}); ok {
-			results[i] = platformclientv2.Supportcentercustommessage{
-				DefaultValue: platformclientv2.String(message["default_value"].(string)),
-				VarType:      platformclientv2.String(message["type"].(string)),
-			}
-		}
-	}
-
-	return &results
-}
-
-func readModuleSettings(settings []interface{}) *[]platformclientv2.Supportcentermodulesetting {
-	if settings == nil || len(settings) < 1 {
-		return nil
-	}
-
-	ret := make([]platformclientv2.Supportcentermodulesetting, len(settings))
-	for i, setting := range settings {
-		settingMap, ok := setting.(map[string]interface{})
-		if !ok {
-			continue
-		}
-
-		moduleSetting := platformclientv2.Supportcentermodulesetting{
-			VarType: resourcedata.GetNillableValueFromMap[string](settingMap, "type"),
-			Enabled: resourcedata.GetNillableValueFromMap[bool](settingMap, "enabled"),
-		}
-
-		if compactModActive, ok := settingMap["compact_category_module_template_active"].(bool); ok {
-			moduleSetting.CompactCategoryModuleTemplate = &platformclientv2.Supportcentercompactcategorymoduletemplate{
-				Active: &compactModActive,
-			}
-		}
-
-		if detailedModTempArr, ok := settingMap["detailed_category_module_template"].([]interface{}); ok && len(detailedModTempArr) > 0 {
-			detailedModTemp := detailedModTempArr[0].(map[string]interface{})
-			moduleSetting.DetailedCategoryModuleTemplate = &platformclientv2.Supportcenterdetailedcategorymoduletemplate{
-				Active: platformclientv2.Bool(detailedModTemp["active"].(bool)),
-				Sidebar: &platformclientv2.Supportcenterdetailedcategorymodulesidebar{
-					Enabled: platformclientv2.Bool(detailedModTemp["sidebar_enabled"].(bool)),
-				},
-			}
-		}
-
-		ret[i] = moduleSetting
-	}
-
-	return &ret
-}
-
-func readScreens(screens []interface{}) *[]platformclientv2.Supportcenterscreen {
-	if screens == nil || len(screens) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Supportcenterscreen, len(screens))
-	for i, value := range screens {
-		if screen, ok := value.(map[string]interface{}); ok {
-			results[i] = platformclientv2.Supportcenterscreen{
-				VarType:        platformclientv2.String(screen["type"].(string)),
-				ModuleSettings: readModuleSettings(screen["module_settings"].([]interface{})),
-			}
-		}
-	}
-
-	return &results
-}
-
-func readEnabledCategories(categories []interface{}) *[]platformclientv2.Supportcentercategory {
-	if categories == nil || len(categories) < 1 {
-		return nil
-	}
-
-	results := make([]platformclientv2.Supportcentercategory, len(categories))
-	for i, value := range categories {
-		category, ok := value.(map[string]interface{})
-		if !ok {
-			continue
-		}
-
-		scCategory := platformclientv2.Supportcentercategory{
-			Id: platformclientv2.String(category["category_id"].(string)),
-		}
-
-		if imageUri, ok := category["image_uri"].(string); ok {
-			scCategory.Image = &platformclientv2.Supportcenterimage{
-				Source: &platformclientv2.Supportcenterimagesource{
-					DefaultUrl: &imageUri,
-				},
-			}
-		}
-
-		results[i] = scCategory
-	}
-
-	return &results
-}
-
-func readSupportCenterHeroStyle(styles []interface{}) *platformclientv2.Supportcenterherostyle {
-	if styles == nil || len(styles) < 1 {
-		return nil
-	}
-
-	style := styles[0].(map[string]interface{})
-	heroStyle := &platformclientv2.Supportcenterherostyle{
-		BackgroundColor: platformclientv2.String(style["background_color"].(string)),
-		TextColor:       platformclientv2.String(style["text_color"].(string)),
-	}
-
-	if imageUri, ok := style["image_uri"].(string); ok {
-		heroStyle.Image = &platformclientv2.Supportcenterimage{
-			Source: &platformclientv2.Supportcenterimagesource{
-				DefaultUrl: &imageUri,
-			},
-		}
-	}
-
-	return heroStyle
-}
-
-func readSupportCenterGlobalStyle(styles []interface{}) *platformclientv2.Supportcenterglobalstyle {
-	if styles == nil || len(styles) < 1 {
-		return nil
-	}
-
-	style := styles[0].(map[string]interface{})
-	globalStyle := &platformclientv2.Supportcenterglobalstyle{
-		BackgroundColor:   platformclientv2.String(style["background_color"].(string)),
-		PrimaryColor:      platformclientv2.String(style["primary_color"].(string)),
-		PrimaryColorDark:  platformclientv2.String(style["primary_color_dark"].(string)),
-		PrimaryColorLight: platformclientv2.String(style["primary_color_light"].(string)),
-		TextColor:         platformclientv2.String(style["text_color"].(string)),
-		FontFamily:        platformclientv2.String(style["font_family"].(string)),
-	}
-
-	return globalStyle
-}
-
-func readStyleSettings(styles []interface{}) *platformclientv2.Supportcenterstylesetting {
-	if styles == nil || len(styles) < 1 {
-		return nil
-	}
-
-	style := styles[0].(map[string]interface{})
-	styleSetting := &platformclientv2.Supportcenterstylesetting{
-		HeroStyle:   readSupportCenterHeroStyle(style["hero_style_setting"].([]interface{})),
-		GlobalStyle: readSupportCenterGlobalStyle(style["global_style_setting"].([]interface{})),
-	}
-
-	return styleSetting
-}
-
-func readMessengerSettings(d *schema.ResourceData) *platformclientv2.Messengersettings {
-	value, ok := d.GetOk("messenger")
-	if !ok {
-		return nil
-	}
-
-	cfgs := value.([]interface{})
-	if len(cfgs) < 1 {
-		return nil
-	}
-
-	cfg := cfgs[0].(map[string]interface{})
-	enabled, _ := cfg["enabled"].(bool)
-	messengerSettings := &platformclientv2.Messengersettings{
-		Enabled: &enabled,
-	}
-
-	if styles, ok := cfg["styles"].([]interface{}); ok && len(styles) > 0 {
-		style := styles[0].(map[string]interface{})
-		if primaryColor, ok := style["primary_color"].(string); ok {
-			messengerSettings.Styles = &platformclientv2.Messengerstyles{
-				PrimaryColor: &primaryColor,
-			}
-		}
-	}
-
-	if launchers, ok := cfg["launcher_button"].([]interface{}); ok && len(launchers) > 0 {
-		launcher := launchers[0].(map[string]interface{})
-		if visibility, ok := launcher["visibility"].(string); ok {
-			messengerSettings.LauncherButton = &platformclientv2.Launcherbuttonsettings{
-				Visibility: &visibility,
-			}
-		}
-	}
-
-	if screens, ok := cfg["home_screen"].([]interface{}); ok && len(screens) > 0 {
-		if screen, ok := screens[0].(map[string]interface{}); ok {
-			enabled, enabledOk := screen["enabled"].(bool)
-			logoUrl, logoUrlOk := screen["logo_url"].(string)
-
-			if enabledOk && logoUrlOk {
-				messengerSettings.HomeScreen = &platformclientv2.Messengerhomescreen{
-					Enabled: &enabled,
-					LogoUrl: &logoUrl,
-				}
-			}
-		}
-	}
-
-	if fileUploads, ok := cfg["file_upload"].([]interface{}); ok && len(fileUploads) > 0 {
-		fileUpload := fileUploads[0].(map[string]interface{})
-		if modesCfg, ok := fileUpload["mode"].([]interface{}); ok && len(modesCfg) > 0 {
-			modes := make([]platformclientv2.Fileuploadmode, len(modesCfg))
-			for i, modeCfg := range modesCfg {
-				if mode, ok := modeCfg.(map[string]interface{}); ok {
-					maxFileSize := mode["max_file_size_kb"].(int)
-					fileTypes := lists.InterfaceListToStrings(mode["file_types"].([]interface{}))
-					modes[i] = platformclientv2.Fileuploadmode{
-						FileTypes:     &fileTypes,
-						MaxFileSizeKB: &maxFileSize,
-					}
-				}
-			}
-
-			if len(modes) > 0 {
-				messengerSettings.FileUpload = &platformclientv2.Fileuploadsettings{
-					Modes: &modes,
-				}
-			}
-		}
-	}
-
-	return messengerSettings
-}
-
-func readCobrowseSettings(d *schema.ResourceData) *platformclientv2.Cobrowsesettings {
-	value, ok := d.GetOk("cobrowse")
-	if !ok {
-		return nil
-	}
-
-	cfgs := value.([]interface{})
-	if len(cfgs) < 1 {
-		return nil
-	}
-
-	cfg := cfgs[0].(map[string]interface{})
-
-	enabled, _ := cfg["enabled"].(bool)
-	allowAgentControl, _ := cfg["allow_agent_control"].(bool)
-	channels := lists.InterfaceListToStrings(cfg["channels"].([]interface{}))
-	maskSelectors := lists.InterfaceListToStrings(cfg["mask_selectors"].([]interface{}))
-	readonlySelectors := lists.InterfaceListToStrings(cfg["readonly_selectors"].([]interface{}))
-
-	return &platformclientv2.Cobrowsesettings{
-		Enabled:           &enabled,
-		AllowAgentControl: &allowAgentControl,
-		Channels:          &channels,
-		MaskSelectors:     &maskSelectors,
-		ReadonlySelectors: &readonlySelectors,
-	}
-}
-
-func readJourneySettings(d *schema.ResourceData) *platformclientv2.Journeyeventssettings {
-	value, ok := d.GetOk("journey_events")
-	if !ok {
-		return nil
-	}
-
-	cfgs := value.([]interface{})
-	if len(cfgs) < 1 {
-		return nil
-	}
-
-	cfg := cfgs[0].(map[string]interface{})
-	enabled, _ := cfg["enabled"].(bool)
-	journeySettings := &platformclientv2.Journeyeventssettings{
-		Enabled: &enabled,
-	}
-
-	excludedQueryParams := lists.InterfaceListToStrings(cfg["excluded_query_parameters"].([]interface{}))
-	journeySettings.ExcludedQueryParameters = &excludedQueryParams
-
-	if keepUrlFragment, ok := cfg["should_keep_url_fragment"].(bool); ok && keepUrlFragment {
-		journeySettings.ShouldKeepUrlFragment = &keepUrlFragment
-	}
-
-	searchQueryParameters := lists.InterfaceListToStrings(cfg["search_query_parameters"].([]interface{}))
-	journeySettings.SearchQueryParameters = &searchQueryParameters
-
-	pageviewConfig := cfg["pageview_config"]
-	if value, ok := pageviewConfig.(string); ok {
-		if value != "" {
-			journeySettings.PageviewConfig = &value
-		}
-	}
-
-	if clickEvents := readSelectorEventTriggers(cfg["click_event"].([]interface{})); clickEvents != nil {
-		journeySettings.ClickEvents = clickEvents
-	}
-
-	if formsTrackEvents := readFormsTrackTriggers(cfg["form_track_event"].([]interface{})); formsTrackEvents != nil {
-		journeySettings.FormsTrackEvents = formsTrackEvents
-	}
-
-	if idleEvents := readIdleEventTriggers(cfg["idle_event"].([]interface{})); idleEvents != nil {
-		journeySettings.IdleEvents = idleEvents
-	}
-
-	if inViewportEvents := readSelectorEventTriggers(cfg["in_viewport_event"].([]interface{})); inViewportEvents != nil {
-		journeySettings.InViewportEvents = inViewportEvents
-	}
-
-	if scrollDepthEvents := readScrollPercentageEventTriggers(cfg["scroll_depth_event"].([]interface{})); scrollDepthEvents != nil {
-		journeySettings.ScrollDepthEvents = scrollDepthEvents
-	}
-
-	return journeySettings
-}
-
-func readSupportCenterSettings(d *schema.ResourceData) *platformclientv2.Supportcentersettings {
-	value, ok := d.GetOk("support_center")
-	if !ok {
-		return nil
-	}
-
-	cfgs := value.([]interface{})
-	if len(cfgs) < 1 {
-		return nil
-	}
-
-	cfg := cfgs[0].(map[string]interface{})
-	supportCenterSettings := &platformclientv2.Supportcentersettings{
-		Enabled:           platformclientv2.Bool(cfg["enabled"].(bool)),
-		EnabledCategories: readEnabledCategories(cfg["enabled_categories"].([]interface{})),
-		StyleSetting:      readStyleSettings(cfg["style_setting"].([]interface{})),
-		Feedback: &platformclientv2.Supportcenterfeedbacksettings{
-			Enabled: platformclientv2.Bool(cfg["feedback_enabled"].(bool)),
-		},
-		RouterType: resourcedata.GetNillableNonZeroValueFromMap[string](cfg, "router_type"),
-	}
-
-	if kbId, ok := cfg["knowledge_base_id"].(string); ok && kbId != "" {
-		supportCenterSettings.KnowledgeBase = &platformclientv2.Addressableentityref{
-			Id: &kbId,
-		}
-	}
-
-	if customMessages := readCustomMessages(cfg["custom_messages"].([]interface{})); customMessages != nil {
-		supportCenterSettings.CustomMessages = customMessages
-	}
-
-	if screens := readScreens(cfg["screens"].([]interface{})); screens != nil {
-		supportCenterSettings.Screens = screens
-	}
-
-	return supportCenterSettings
-}
-
-// featureNotImplemented checks the response object to find out if the request failed because a feature is not yet
-// implemented in the org that it was ran against. If true, we can pass back the field name and give more context
-// in the final error message.
-func featureNotImplemented(response *platformclientv2.APIResponse) (bool, string) {
-	if response.Error == nil || response.Error.Details == nil || len(response.Error.Details) == 0 {
-		return false, ""
-	}
-	for _, err := range response.Error.Details {
-		if err.FieldName == nil {
-			continue
-		}
-		if strings.Contains(*err.ErrorCode, "feature is not yet implemented") {
-			return true, *err.FieldName
-		}
-	}
-	return false, ""
-}
-
-func validateConfigurationStatusChange(k, old, new string, d *schema.ResourceData) bool {
-	// Configs start in a pending status and may not transition to active or error before we retrieve the state, so allow
-	// the status to change from pending to something less ephemeral
-	return old == "Pending"
-}
-
-func flattenMessengerSettings(messengerSettings *platformclientv2.Messengersettings) []interface{} {
-	if messengerSettings == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"enabled":         messengerSettings.Enabled,
-		"styles":          flattenStyles(messengerSettings.Styles),
-		"launcher_button": flattenLauncherButton(messengerSettings.LauncherButton),
-		"home_screen":     flattenHomeScreen(messengerSettings.HomeScreen),
-		"file_upload":     flattenFileUpload(messengerSettings.FileUpload),
-	}}
-}
-
-func flattenCobrowseSettings(cobrowseSettings *platformclientv2.Cobrowsesettings) []interface{} {
-	if cobrowseSettings == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"enabled":             cobrowseSettings.Enabled,
-		"allow_agent_control": cobrowseSettings.AllowAgentControl,
-		"channels":            cobrowseSettings.Channels,
-		"mask_selectors":      cobrowseSettings.MaskSelectors,
-		"readonly_selectors":  cobrowseSettings.ReadonlySelectors,
-	}}
-}
-
-func flattenStyles(styles *platformclientv2.Messengerstyles) []interface{} {
-	if styles == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"primary_color": styles.PrimaryColor,
-	}}
-}
-
-func flattenLauncherButton(settings *platformclientv2.Launcherbuttonsettings) []interface{} {
-	if settings == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"visibility": settings.Visibility,
-	}}
-}
-
-func flattenHomeScreen(settings *platformclientv2.Messengerhomescreen) []interface{} {
-	if settings == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"enabled":  settings.Enabled,
-		"logo_url": settings.LogoUrl,
-	}}
-}
-
-func flattenFileUpload(settings *platformclientv2.Fileuploadsettings) []interface{} {
-	if settings == nil || settings.Modes == nil || len(*settings.Modes) < 1 {
-		return nil
-	}
-
-	modes := make([]map[string]interface{}, len(*settings.Modes))
-	for i, mode := range *settings.Modes {
-		modes[i] = map[string]interface{}{
-			"file_types":       *mode.FileTypes,
-			"max_file_size_kb": *mode.MaxFileSizeKB,
-		}
-	}
-
-	return []interface{}{map[string]interface{}{
-		"mode": modes,
-	}}
-}
-
-func flattenJourneyEvents(journeyEvents *platformclientv2.Journeyeventssettings) []interface{} {
-	if journeyEvents == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"enabled":                   journeyEvents.Enabled,
-		"excluded_query_parameters": journeyEvents.ExcludedQueryParameters,
-		"should_keep_url_fragment":  journeyEvents.ShouldKeepUrlFragment,
-		"search_query_parameters":   journeyEvents.SearchQueryParameters,
-		"pageview_config":           journeyEvents.PageviewConfig,
-		"click_event":               flattenSelectorEventTriggers(journeyEvents.ClickEvents),
-		"form_track_event":          flattenFormsTrackTriggers(journeyEvents.FormsTrackEvents),
-		"idle_event":                flattenIdleEventTriggers(journeyEvents.IdleEvents),
-		"in_viewport_event":         flattenSelectorEventTriggers(journeyEvents.InViewportEvents),
-		"scroll_depth_event":        flattenScrollPercentageEventTriggers(journeyEvents.ScrollDepthEvents),
-	}}
-}
-
-func flattenSelectorEventTriggers(triggers *[]platformclientv2.Selectoreventtrigger) []interface{} {
-	if triggers == nil || len(*triggers) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*triggers))
-	for i, trigger := range *triggers {
-		result[i] = map[string]interface{}{
-			"selector":   trigger.Selector,
-			"event_name": trigger.EventName,
-		}
-	}
-	return result
-}
-
-func flattenFormsTrackTriggers(triggers *[]platformclientv2.Formstracktrigger) []interface{} {
-	if triggers == nil || len(*triggers) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*triggers))
-	for i, trigger := range *triggers {
-		result[i] = map[string]interface{}{
-			"selector":                     trigger.Selector,
-			"form_name":                    trigger.FormName,
-			"capture_data_on_form_abandon": trigger.CaptureDataOnFormAbandon,
-			"capture_data_on_form_submit":  trigger.CaptureDataOnFormSubmit,
-		}
-	}
-	return result
-}
-
-func flattenIdleEventTriggers(triggers *[]platformclientv2.Idleeventtrigger) []interface{} {
-	if triggers == nil || len(*triggers) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*triggers))
-	for i, trigger := range *triggers {
-		result[i] = map[string]interface{}{
-			"event_name":         trigger.EventName,
-			"idle_after_seconds": trigger.IdleAfterSeconds,
-		}
-	}
-	return result
-}
-
-func flattenScrollPercentageEventTriggers(triggers *[]platformclientv2.Scrollpercentageeventtrigger) []interface{} {
-	if triggers == nil || len(*triggers) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*triggers))
-	for i, trigger := range *triggers {
-		result[i] = map[string]interface{}{
-			"event_name": trigger.EventName,
-			"percentage": trigger.Percentage,
-		}
-	}
-	return result
-}
-
-func flattenSupportCenterSettings(supportCenterSettings *platformclientv2.Supportcentersettings) []interface{} {
-	if supportCenterSettings == nil {
-		return nil
-	}
-
-	settingsMap := map[string]interface{}{
-		"enabled":            supportCenterSettings.Enabled,
-		"custom_messages":    flattenCustomMessages(supportCenterSettings.CustomMessages),
-		"router_type":        supportCenterSettings.RouterType,
-		"screens":            flattenScreens(supportCenterSettings.Screens),
-		"enabled_categories": flattenEnabledCategories(supportCenterSettings.EnabledCategories),
-	}
-
-	if supportCenterSettings.KnowledgeBase != nil {
-		settingsMap["knowledge_base_id"] = supportCenterSettings.KnowledgeBase.Id
-	}
-
-	if supportCenterSettings.Feedback != nil {
-		settingsMap["feedback_enabled"] = supportCenterSettings.Feedback.Enabled
-	}
-
-	if supportCenterSettings.StyleSetting != nil {
-		settingsMap["style_setting"] = flattenStyleSettings(supportCenterSettings.StyleSetting)
-	}
-
-	return []interface{}{settingsMap}
-}
-
-func flattenCustomMessages(messages *[]platformclientv2.Supportcentercustommessage) []interface{} {
-	if messages == nil || len(*messages) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*messages))
-	for i, message := range *messages {
-		result[i] = map[string]interface{}{
-			"default_value": message.DefaultValue,
-			"type":          message.VarType,
-		}
-	}
-	return result
-}
-
-func flattenScreens(screens *[]platformclientv2.Supportcenterscreen) []interface{} {
-	if screens == nil || len(*screens) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*screens))
-	for i, screen := range *screens {
-		result[i] = map[string]interface{}{
-			"type":            screen.VarType,
-			"module_settings": flattenModuleSettings(screen.ModuleSettings),
-		}
-	}
-	return result
-}
-
-func flattenModuleSettings(settings *[]platformclientv2.Supportcentermodulesetting) []interface{} {
-	if settings == nil || len(*settings) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*settings))
-	for i, setting := range *settings {
-		settingMap := map[string]interface{}{
-			"type":    setting.VarType,
-			"enabled": setting.Enabled,
-		}
-
-		if setting.CompactCategoryModuleTemplate != nil {
-			settingMap["compact_category_module_template_active"] = setting.CompactCategoryModuleTemplate.Active
-		}
-
-		if setting.DetailedCategoryModuleTemplate != nil {
-			settingMap["detailed_category_module_template"] = []interface{}{
-				map[string]interface{}{
-					"active":          setting.DetailedCategoryModuleTemplate.Active,
-					"sidebar_enabled": setting.DetailedCategoryModuleTemplate.Sidebar.Enabled,
-				},
-			}
-		}
-
-		result[i] = settingMap
-	}
-	return result
-}
-
-func flattenEnabledCategories(categories *[]platformclientv2.Supportcentercategory) []interface{} {
-	if categories == nil || len(*categories) < 1 {
-		return nil
-	}
-
-	result := make([]interface{}, len(*categories))
-	for i, category := range *categories {
-		categoryMap := map[string]interface{}{
-			"category_id": category.Id,
-		}
-
-		if category.Image != nil && category.Image.Source != nil && category.Image.Source.DefaultUrl != nil {
-			categoryMap["image_uri"] = *category.Image.Source.DefaultUrl
-		}
-
-		result[i] = categoryMap
-	}
-	return result
-}
-
-func flattenStyleSettings(style *platformclientv2.Supportcenterstylesetting) []interface{} {
-	if style == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"hero_style_setting":   flattenHeroStyle(style.HeroStyle),
-		"global_style_setting": flattenGlobalStyle(style.GlobalStyle),
-	}}
-}
-
-func flattenHeroStyle(style *platformclientv2.Supportcenterherostyle) []interface{} {
-	if style == nil {
-		return nil
-	}
-
-	styleMap := map[string]interface{}{
-		"background_color": style.BackgroundColor,
-		"text_color":       style.TextColor,
-	}
-
-	if style.Image != nil && style.Image.Source != nil && style.Image.Source.DefaultUrl != nil {
-		styleMap["image_uri"] = *style.Image.Source.DefaultUrl
-	}
-
-	return []interface{}{styleMap}
-}
-
-func flattenGlobalStyle(style *platformclientv2.Supportcenterglobalstyle) []interface{} {
-	if style == nil {
-		return nil
-	}
-
-	return []interface{}{map[string]interface{}{
-		"background_color":    style.BackgroundColor,
-		"primary_color":       style.PrimaryColor,
-		"primary_color_dark":  style.PrimaryColorDark,
-		"primary_color_light": style.PrimaryColorLight,
-		"text_color":          style.TextColor,
-		"font_family":         style.FontFamily,
-	}}
-}
-
-func readWebDeploymentConfigurationFromResourceData(d *schema.ResourceData) (string, *platformclientv2.Webdeploymentconfigurationversion) {
-	name := d.Get("name").(string)
-	languages := lists.InterfaceListToStrings(d.Get("languages").([]interface{}))
-	defaultLanguage := d.Get("default_language").(string)
-
-	inputCfg := &platformclientv2.Webdeploymentconfigurationversion{
-		Name:            &name,
-		Languages:       &languages,
-		DefaultLanguage: &defaultLanguage,
-	}
-
-	description, ok := d.Get("description").(string)
-	if ok {
-		inputCfg.Description = &description
-	}
-
-	messengerSettings := readMessengerSettings(d)
-	if messengerSettings != nil {
-		inputCfg.Messenger = messengerSettings
-	}
-
-	cobrowseSettings := readCobrowseSettings(d)
-	if cobrowseSettings != nil {
-		inputCfg.Cobrowse = cobrowseSettings
-	}
-
-	journeySettings := readJourneySettings(d)
-	if journeySettings != nil {
-		inputCfg.JourneyEvents = journeySettings
-	}
-
-	supportCenterSettings := readSupportCenterSettings(d)
-	if supportCenterSettings != nil {
-		inputCfg.SupportCenter = supportCenterSettings
-	}
-
-	return name, inputCfg
-}
